Oops; gotta be careful with those lists of integers.
git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@1731 edc67d0a-ffea-dc11-983e-003048321218

+++ src/main/java/com/threerings/tudey/server/util/Pathfinder.java
@@ -518,7 +518,7 @@ public class Pathfinder
                                 if (list == null) {
                                     continue;
                                 }
-                                list.remove(flags);
+                                list.remove((Integer)flags);
                                 int nsize = list.size();
                                 if (nsize == 0) {
                                     _actorFlagLists.remove(coord);
